refactor: declaring declarative functions and typing as functional component
