+++ src/components/ProfileSettings/Sidebar/index.tsx
@@ -1,12 +1,11 @@
-import Image from 'next/image'
+import { FunctionComponent } from 'react'
 import Link from 'next/link'
-import UploadAvatar from './UploadAvatar'
 import EditUserContainer from '../EditUserContainer'
 
-export default function Sidebar() {
+const Sidebar: FunctionComponent = () => {
   return (
     <section className="h-[18.350rem] flex justify-center gap-6">
-      <div className="">
+      <div className="w-[220px]">
         <h2 className="font-bold text-xl">Em Geral</h2>
         <div className="flex flex-col pt-[2.8125rem]">
           <Link href="" className="font-bold text-[0.8125rem] text-default">
@@ -18,3 +17,5 @@ export default function Sidebar() {
     </section>
   )
 }
+
+export default Sidebar
